requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.0.0
numpy>=1.24.0
openai>=1.0.0
finnhub-python>=2.4.0
yfinance>=0.2.0
ta>=0.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
schedule>=1.2.0
aiohttp>=3.8.0